Add "wait for moderator" feature
This feature allows the use of the "w" flag when joining a conference
to signify that normal callers to the conference should have MOH played
until someone marked as a moderator ("M" flag) joins the conference.

It differs from the builtin "H" feature in two primary ways.  First,
multiple callers can be in the conference only hearing the MOH prior to
the arrival of the moderator.  Secondly, if the moderator is the first
to join, he will not hear the MOH, but will instead be dropped directly
into the conference, albeit all by himself.